Use double instead of float

+++ semester 3/programming languages/09-07-2018/2.c
@@ -3,30 +3,30 @@
 #include <math.h>
 
 int main(int argc, char *argv[]) {
-	float A, B, C, D;
+	double A, B, C, D;
 	puts("Ax^2+Bx+C=0");
 
 	printf("Enter A: ");
-	scanf("%f", &A);
+	scanf("%lf", &A);
 
 	printf("Enter B: ");
-	scanf("%f", &B);
+	scanf("%lf", &B);
 
 	printf("Enter C: ");
-	scanf("%f", &C);
+	scanf("%lf", &C);
 
 	D = B*B - 4*A*C;
 	puts("Answers:");
 
 	if (D < 0) {
-		float re = -B/2/A;
-		float im = sqrt(fabs(D))/2/A;
+		double re = -B/2/A;
+		double im = sqrt(fabs(D))/2/A;
 		printf("x1: %f+%fi,\nx2: %f-%fi.\n", re, im, re, im);
 		return 0;
 	}
 
-	float x1 = (-B + sqrt(D))/2/A;
-	float x2 = (-B - sqrt(D))/2/A;
+	double x1 = (-B + sqrt(D))/2/A;
+	double x2 = (-B - sqrt(D))/2/A;
 
 	printf("x1: %f,\nx2: %f.\n", x1, x2);
